// deno-api/md/types/mdast.d.ts
export type Content =
  | TopLevelContent
  | ListContent
  | TableContent
  | RowContent
  | PhrasingContent;
// ...
export interface Parent extends UnistParent {
  children: Content[];
}

export interface Literal extends UnistLiteral {
  value: string;
}

export interface Root extends Parent {
  type: "root";
}